clean up : component
