+++ src/ecs_elements/components/menus.rs
@@ -1,0 +1,16 @@
+use bevy::prelude::*;
+
+#[derive(Component)]
+pub struct OnScreenMenu;
+
+
+#[derive(Component)]
+pub struct NineSlice;
+
+
+#[derive(Component, PartialEq, Clone, Copy)]
+pub enum MainMenuOptions {
+    StartGame,
+    LoadGame,
+    Quit
+}
