+++ src/ecs_elements/components/text.rs
@@ -1,0 +1,5 @@
+use bevy::prelude::*;
+
+#[derive(Component)]
+pub struct AsciiText;
+
