+++ src/ecs_elements/components/tiles.rs
@@ -1,0 +1,11 @@
+use bevy::prelude::*;
+
+
+#[derive(Component)]
+pub struct TileCollider;
+
+#[derive(Component)]
+pub struct TileExit;
+
+#[derive(Component)]
+pub struct GameMap;
