+++ src/ecs_elements/mod.rs
@@ -1,0 +1,3 @@
+pub mod components;
+
+pub use components::*;
