+++ src/game/npc.rs
@@ -6,9 +6,7 @@ use pathfinding::prelude::astar;
 
 use crate::{
     GameState, despawn_screen, TILE_SIZE,
-    game::{
-        Player, Stats, Npc, Monster,
-    },
+    ecs_elements::components::{Player, Stats, Npc, Monster, Pathfinding, MoveTo},
     commons::tile_collision_check,
     map_builders::{
         pathfinding::{Position, world_to_grid_position, grid_to_world_position},
@@ -35,28 +33,6 @@ impl Plugin for NpcPlugin{
             .insert_resource(FixedTime::new_from_secs(FIXED_TIMESTEP))
             ;         
     }
-}
-
-
-
-#[derive(Component)]
-pub struct DisplayedPath;
-
-#[derive(Component)]
-pub struct Pathfinding{
-    pub start: Position,
-    pub goal: Position,
-    pub path: Vec<Position>,
-    pub step: usize,
-    pub dirty: bool,    //Si True, verifie la position vs Step Destination pour savoir si chemin atteint et next ordre de mouvement necessaire.
-    pub debug: bool,
-}
-
-#[derive(Component)]
-pub struct MoveTo{
-    pub x: f32,
-    pub y: f32,
-    pub destination: Position
 }
