+++ src/game/player.rs
@@ -3,7 +3,8 @@ use bevy::prelude::*;
 use crate::{
     commons::{tile_collision_check},
     TILE_SIZE, GameState, despawn_screen,
-    game::{Player, Stats, TileCollider, TileExit, ShouldSave}
+    game::{ShouldSave},
+    ecs_elements::components::{TileCollider, TileExit, Player, Stats, }
 };
