+++ src/menus/mainmenu.rs
@@ -1,26 +1,19 @@
 use bevy::{prelude::*, app::AppExit};
 
 use crate::{
-    despawn_screen, AppState, GameState, HEIGHT, CHAR_SIZE, RESOLUTION,
-    ascii::{spawn_ascii_text, AsciiSheet, NineSliceIndices, spawn_nine_slice, NineSlice},
+    despawn_screen, AppState, GameState, HEIGHT, CHAR_SIZE, 
+    ascii::{spawn_ascii_text, AsciiSheet, NineSliceIndices, spawn_nine_slice, },
     save_load_system::has_save_file,
-    //SHOW_MAPGEN_VISUALIZER,
+    ecs_elements::components::{MainMenuOptions, OnScreenMenu, NineSlice,},
 };
 
 
 
 // ENUMS
 
-#[derive(Component, PartialEq, Clone, Copy)]
-pub enum MainMenuOptions {
-    StartGame,
-    LoadGame,
-    Quit
-}
 
-// COMPONENTS & RESOURCES
-#[derive(Component)]
-pub struct OnScreenMenu;
+
+
 
 pub const MAIN_MENU_OPTIONS_COUNT: isize = 3;  //Necessaire pour la selection d'une option dans l'input.
 
@@ -200,7 +193,7 @@ fn spawn_title(
     let middle_upper = HEIGHT / 4.0;
     let img_width = 384.0;
     //let img_height = 48.0;
-    let img_mid_x = (HEIGHT * RESOLUTION / 4.0) - (img_width / 2.0);
+    //let img_mid_x = (HEIGHT * RESOLUTION / 4.0) - (img_width / 2.0);
 
     let title_drop = commands.spawn(SpriteBundle {
         texture: asset_server.load("title/shadowrun_title_alone.png"),    //asset_server.load("temp_tiles/Sewers_wall.png"),
